Created OpportunityBookmark component
